ravb_streaming: auto detect ravb network interface
In function ravb_streaming_init() network device is referenced by
dev_get_by_name(). This might cause a problem when the ravb capable
interface is not known to the user. Since this module is rcar specific
code, so it is more effecient to auto detect ravb rcar interface
based on ravb_streaming_match_table if no interface specified in
interface parameter.

Signed-off-by: Abdalla Ashry <aashry@de.adit-jv.com>
Signed-off-by: Andrew Gabbasov <andrew_gabbasov@mentor.com>
Signed-off-by: Masaru Nagai <masaru.nagai.vx@renesas.com>