ravb_proc: Remove redundant code in init function
Since "struct device" member "dev" is embedded into "struct net_device"
as a whole structure body, the pointer to its beginning can not be NULL
in a non-NULL enclosing structure. So, if "ndev" is not NULL,
"&ndev->dev" can't be NULL too, and extra check is not needed.

Since removing the above check leaves the only case of "ndev_dev" variable
usage, remove the variable too.

Also, the result of "to_stp" macro (which is the redefinition of
"container_of" macro) can't be NULL, so the check is not needed.

Signed-off-by: Andrew Gabbasov <andrew_gabbasov@mentor.com>
Signed-off-by: Masaru Nagai <masaru.nagai.vx@renesas.com>